requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
distro==1.8.0
Flask==3.0.0
Flask-WTF==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
numpy==1.26.3
openai==1.6.0
opencv-python==4.9.0.80
packaging==23.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pypdfium2==4.25.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
sympy==1.12
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.2
yarl==1.9.4
